@@ -16,16 +16,19 @@
     "psr-4": {"Padam87\\GlsBundle\\": ""}
   },
   "require": {
-    "php": "~5.6|~7.0",
+    "php": "^7.4 || ^8.0",
     "ext-soap": "*",
-    "symfony/config": "^2.8 || ^3.0 || ^4.0",
-    "symfony/dependency-injection": "^2.8.3 || ^3.0.3 || ^4.0",
-    "symfony/options-resolver": "^2.8 || ^3.0 || ^4.0",
-    "php-http/httplug-bundle": "^1.7"
-  },
-  "require-dev": {
-    "php-http/guzzle6-adapter": "^1.1"
+    "symfony/config": "^4.0 || ^5.0 || ^6.0",
+    "symfony/dependency-injection": "^4.0 || ^5.0 || ^6.0",
+    "symfony/form": "^4.0 || ^5.0 || ^6.0",
+    "symfony/options-resolver": "^4.0 || ^5.0 || ^6.0",
+    "symfony/validator": "^4.0 || ^5.0 || ^6.0",
+    "doctrine/common": "^2.0 || ^3.0",
+    "doctrine/collections": "^2.1"
   },
   "suggest": {
+  },
+  "require-dev": {
+    "rector/rector": "^0.8.13"
   }
 }
