@@ -2,124 +2,140 @@
 
 namespace Padam87\GlsBundle\Service;
 
+use Padam87\GlsBundle\Dto\Message\PrintLabels;
+use Padam87\GlsBundle\Dto\Request\AbstractRequest;
+use Padam87\GlsBundle\Dto\Request\GetParcelStatusesRequest;
+use Padam87\GlsBundle\Dto\Request\GetPrintedLabelsRequest;
+use Padam87\GlsBundle\Dto\Request\PrepareLabelsRequest;
+use Padam87\GlsBundle\Dto\Request\PrintLabelsRequest;
+use Padam87\GlsBundle\Dto\Response\GetParcelStatusesResponse;
+use Padam87\GlsBundle\Dto\Response\GetPrintedLabelsResponse;
+use Padam87\GlsBundle\Dto\Response\PrepareLabelsResponse;
+use Padam87\GlsBundle\Dto\Response\PrintLabelsResponse;
+use Padam87\GlsBundle\Model\Address;
+use Padam87\GlsBundle\Model\Collection;
+use Padam87\GlsBundle\Model\ErrorInfo;
+use Padam87\GlsBundle\Model\Parcel;
+use Padam87\GlsBundle\Model\ParcelInfo;
+use Padam87\GlsBundle\Model\ParcelStatus;
+use Padam87\GlsBundle\Model\PrintLabelsInfo;
+use Padam87\GlsBundle\Model\Service;
+use Symfony\Component\HttpClient\CurlHttpClient;
+use Symfony\Component\HttpClient\Exception\ClientException;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class ParcelApi
 {
-    private $config = [];
+    private array $config = [];
+    private ?\SoapClient $client = null;
 
     public function setConfig(array $config)
     {
         $this->config = $config;
     }
 
-    public function getClient(): \SoapClient
+    public function getClient()
     {
-        return new \SoapClient($this->config['parcel_wsdl']);
+        if ($this->client === null) {
+            $this->client = new \SoapClient(
+                $this->config['wsdl'],
+                [
+                    'trace' => 1,
+                    'features' => SOAP_SINGLE_ELEMENT_ARRAYS,
+                    'classmap' => [
+                        // Generics can't come soon enough.
+                        'ArrayOfstring' => Collection::class,
+                        'ArrayOfint' => Collection::class,
+                        'ArrayOfParcel' => Collection::class,
+                        'ArrayOfService' => Collection::class,
+                        'ArrayOfParcelInfo' => Collection::class,
+                        'ArrayOfPrintDataInfo' => Collection::class,
+                        'ArrayOfSuccessfullyDeleted' => Collection::class,
+                        'ArrayOfErrorInfo' => Collection::class,
+                        'ArrayOfPrintLabelsInfo' => Collection::class,
+                        'ArrayOfParcelStatus' => Collection::class,
+
+                        // Commented lines because SOAP sets the wrong namespace for some reason.
+
+                        //'Address' => Address::class,
+                        'Parcel' => Parcel::class,
+                        //'Service' => Service::class,
+
+                        'ErrorInfo' => ErrorInfo::class,
+                        'ParcelInfo' => ParcelInfo::class,
+                        'ParcelStatus' => ParcelStatus::class,
+                        'PrintLabelsInfo' => PrintLabelsInfo::class,
+
+                        //'PrintLabelsRequest' => PrintLabelsRequest::class,
+                        'PrintLabelsResponse' => PrintLabelsResponse::class,
+
+                        //'PrepareLabelsRequest' => PrepareLabelsRequest::class,
+                        'PrepareLabelsResponse' => PrepareLabelsResponse::class,
+
+                        //'GetPrintedLabelsRequest' => GetPrintedLabelsRequest::class,
+                        'GetPrintedLabelsResponse' => GetPrintedLabelsResponse::class,
+
+                        //'GetParcelStatusesRequest' => GetParcelStatusesRequest::class,
+                        'GetParcelStatusResponse' => GetParcelStatusesResponse::class,
+                    ]
+                ]
+            );
+        }
+
+        return $this->client;
     }
 
-    public function createParcel($data): array
+    protected function prepareRequest(AbstractRequest $request)
     {
-        $data = $this->resolveData($data);
-
-        return (array) $this->getClient()->__soapCall('printlabel', $data);
-    }
-
-    /**
-     * @param array $data
-     *
-     * @return array
-     */
-    protected function resolveData(array $data)
-    {
-        $resolver = new OptionsResolver();
-
-        $resolver
-            ->setRequired(
-                [
-                    'username',
-                    'password',
-                    'senderid',
-                    'sender_name',
-                    'sender_address',
-                    'sender_city',
-                    'sender_zipcode',
-                    'sender_country',
-                    'consig_name',
-                    'consig_address',
-                    'consig_city',
-                    'consig_zipcode',
-                    'consig_country',
-                    'pcount',
-                    'pickupdate',
-                    'printertemplate',
-                    'printit',
-                    'timestamp',
-                ]
-            )
-            ->setDefaults(
-                [
-                    'sender_contact' => '',
-                    'sender_phone' => '',
-                    'sender_email' => '',
-                    'consig_contact' => '',
-                    'consig_phone' => '',
-                    'consig_email' => '',
-                    'content' => '',
-                    'clientref' => '',
-                    'codamount' => 0.0,
-                    'codref' => '',
-                    'services' => '',
-                    'timestamp' => date('YmdHis'),
-                    'hash' => '',
-                    'customlabel' => '',
-                    'is_autoprint_pdfs' => false,
-                ]
-            )
-            ->setDefaults($this->config['config'])
+        $request
+            ->setUsername($this->config['config']['username'])
+            ->setPassword(hash('sha512', $this->config['config']['password'], true))
         ;
-
-        $order = [
-            'username', 'password', 'senderid',
-            'sender_name', 'sender_address', 'sender_city', 'sender_zipcode', 'sender_country',
-            'sender_contact', 'sender_phone', 'sender_email',
-            'consig_name', 'consig_address', 'consig_city', 'consig_zipcode', 'consig_country',
-            'consig_contact', 'consig_phone', 'consig_email',
-            'pcount', 'pickupdate', 'content', 'clientref', 'codamount', 'codref',
-            'services', 'printertemplate', 'printit', 'timestamp', 'hash', 'customlabel', 'is_autoprint_pdfs',
-        ];
-
-        $data = $resolver->resolve($data);
-        $return = [];
-
-        foreach ($order as $field) {
-            $return[$field] = $data[$field];
-        }
-
-        $return['hash'] = $this->hash($return);
-
-        return $return;
     }
 
-    /**
-     * @param $data
-     *
-     * @return string
-     */
-    protected function hash($data)
+    public function prepareLabels(PrepareLabelsRequest $request): PrepareLabelsResponse
     {
-        $excluded = ['services', 'printertemplate', 'printit', 'timestamp', 'hash', 'customlabel', 'is_autoprint_pdfs'];
+        $this->prepareRequest($request);
 
-        $hashBase = '';
+        $request->getParcelList()->forAll(function ($key, Parcel $parcel) {
+            $parcel->setClientNumber($this->config['config']['senderid']);
+        });
 
-        foreach ($data as $field => $value) {
-            if (in_array($field, $excluded)) {
-                continue;
-            }
+        $response = $this->getClient()->PrepareLabels(['prepareLabelsRequest' => $request]);
 
-            $hashBase .= $data[$field];
-        }
+        return $response->PrepareLabelsResult;
+    }
 
-        return sha1($hashBase);
+    public function getPrintedLabels(GetPrintedLabelsRequest $request): GetPrintedLabelsResponse
+    {
+        $this->prepareRequest($request);
+
+        $response = $this->getClient()->GetPrintedLabels(['getPrintedLabelsRequest' => $request]);
+
+        return $response->GetPrintedLabelsResult;
+    }
+
+    public function printLabels(PrintLabelsRequest $request): PrintLabelsResponse
+    {
+        $this->prepareRequest($request);
+
+        $request->getParcelList()->forAll(function ($key, Parcel $parcel) {
+            $parcel->setClientNumber($this->config['config']['senderid']);
+
+            return true;
+        });
+
+        $response = $this->getClient()->PrintLabels(['printLabelsRequest' => $request]);
+
+        return $response->PrintLabelsResult;
+    }
+
+    public function getParcelStatuses(GetParcelStatusesRequest $request): GetParcelStatusesResponse
+    {
+        $this->prepareRequest($request);
+
+        $response = $this->getClient()->GetParcelStatuses(['getParcelStatusesRequest' => $request]);
+
+        return $response->GetParcelStatusesResult;
     }
 }
