@@ -2,9 +2,9 @@
 
 namespace Padam87\GlsBundle\DependencyInjection;
 
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 use Padam87\GlsBundle\Service\ParcelApi;
-use Padam87\GlsBundle\Service\PodDownloadApi;
-use Padam87\GlsBundle\Service\TrackingApi;
+use Padam87\GlsBundle\Service\ParcelGeneratorInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
@@ -25,16 +25,12 @@ class Padam87GlsExtension extends Extension
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
 
-        $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.yaml');
 
         $definition = $container->getDefinition(ParcelApi::class);
         $definition->addMethodCall('setConfig', [$config]);
 
-        $definition = $container->getDefinition(TrackingApi::class);
-        $definition->addMethodCall('setConfig', [$config]);
-
-        $definition = $container->getDefinition(PodDownloadApi::class);
-        $definition->addMethodCall('setConfig', [$config]);
+        $container->registerForAutoconfiguration(ParcelGeneratorInterface::class)->addTag('padam87_gls.parcel_genarator');
     }
 }
